Reconcile the two flavors of `unit-capi-serialized_queries.cc`.
