+++ test/src/unit-capi-serialized_queries.cc
@@ -118,6 +118,22 @@ struct SerializationFx {
     REQUIRE(loop_num->second > 0);
   }
 
+  static void check_subarray_stats(int dim0_expected, int dim1_expected) {
+    Stats::enable();
+    std::string stats;
+    Stats::dump(&stats);
+    // Note: if these checks fail, use Stats::dump(stdout) to validate counters
+    CHECK(
+        stats.find(
+            "\"Context.StorageManager.subSubarray.add_range_dim_0\": " +
+            std::to_string(dim0_expected)) != std::string::npos);
+    CHECK(
+        stats.find(
+            "\"Context.StorageManager.subSubarray.add_range_dim_1\": " +
+            std::to_string(dim1_expected)) != std::string::npos);
+    Stats::disable();
+  }
+
   static void check_delete_stats(const Query& query) {
     auto stats = ((sm::Reader*)query.ptr()->query_->strategy())->stats();
     REQUIRE(stats != nullptr);
@@ -177,7 +193,9 @@ struct SerializationFx {
 
     Array array(ctx, array_uri, TILEDB_WRITE);
     Query query(ctx, array);
-    query.set_subarray(subarray);
+    Subarray cppapi_subarray(ctx, array);
+    cppapi_subarray.set_subarray(subarray);
+    query.set_subarray(cppapi_subarray);
     query.set_data_buffer("a1", a1);
     query.set_data_buffer("a2", a2);
     query.set_validity_buffer("a2", a2_nullable);
@@ -218,10 +236,10 @@ struct SerializationFx {
 
     Array array(ctx, array_uri, TILEDB_WRITE);
     Query query(ctx, array);
-    Subarray sub(ctx, array);
-    sub.add_range(0, subarray[0], subarray[1]);
-    sub.add_range(1, subarray[2], subarray[3]);
-    query.set_subarray(sub);
+    Subarray cppapi_subarray(ctx, array);
+    cppapi_subarray.add_range(0, subarray[0], subarray[1]);
+    cppapi_subarray.add_range(1, subarray[2], subarray[3]);
+    query.set_subarray(cppapi_subarray);
     query.set_data_buffer("a1", a1);
     query.set_data_buffer("a2", a2);
     query.set_validity_buffer("a2", a2_nullable);
@@ -345,6 +363,7 @@ TEST_CASE_METHOD(
     "[query][dense][serialization][rest]") {
   create_array(TILEDB_DENSE);
   auto expected_results = write_dense_array();
+  check_subarray_stats(2, 2);
 
   SECTION("- Read all") {
     Array array(ctx, array_uri, TILEDB_READ);
@@ -356,18 +375,26 @@ TEST_CASE_METHOD(
     std::vector<uint64_t> a3_offsets(1000);
     std::vector<int32_t> subarray = {1, 10, 1, 10};
 
-    query.set_subarray(subarray);
+    Subarray cppapi_subarray(ctx, array);
+    cppapi_subarray.set_subarray(subarray);
+    query.set_subarray(cppapi_subarray);
     query.set_data_buffer("a1", a1);
     query.set_data_buffer("a2", a2);
     query.set_validity_buffer("a2", a2_nullable);
     query.set_data_buffer("a3", a3_data);
     query.set_offsets_buffer("a3", a3_offsets);
 
+    // Check stats before serialization
+    check_subarray_stats(3, 3);
+
     // Submit query
     query.submit();
+    REQUIRE(query.query_status() == Query::Status::COMPLETE);
 
-    // The deserialized query should also include the read stats
+    // Check stats after serialization
+    // #TODO Revisit after Stats serialization is reworked
     check_read_stats(query);
+    check_subarray_stats(5, 5);
 
     auto result_el = query.result_buffer_elements_nullable();
     REQUIRE(std::get<1>(result_el["a1"]) == 100);
@@ -393,7 +420,9 @@ TEST_CASE_METHOD(
     std::vector<uint64_t> a3_offsets(1000);
     std::vector<int32_t> subarray = {1, 10, 1, 10};
 
-    query.set_subarray(subarray);
+    Subarray cppapi_subarray(ctx, array);
+    cppapi_subarray.set_subarray(subarray);
+    query.set_subarray(cppapi_subarray);
     query.set_data_buffer("a1", a1);
     query.set_data_buffer("a2", a2);
     query.set_validity_buffer("a2", a2_nullable);
@@ -407,9 +436,11 @@ TEST_CASE_METHOD(
 
     // Submit query
     query.submit();
+    REQUIRE(query.query_status() == Query::Status::COMPLETE);
 
-    // The deserialized query should also include the read stats
+    // The deserialized query should also include the write stats
     check_read_stats(query);
+    check_subarray_stats(5, 5);
 
     // We expect all cells where `a1` >= `cmp_value` to be filtered
     // out. For the refactored reader, filtered out means the value is
@@ -456,7 +487,9 @@ TEST_CASE_METHOD(
     std::vector<uint64_t> a3_offsets(1000);
     std::vector<int32_t> subarray = {3, 4, 3, 4};
 
-    query.set_subarray(subarray);
+    Subarray cppapi_subarray(ctx, array);
+    cppapi_subarray.set_subarray(subarray);
+    query.set_subarray(cppapi_subarray);
     query.set_data_buffer("a1", a1);
     query.set_data_buffer("a2", a2);
     query.set_validity_buffer("a2", a2_nullable);
@@ -465,8 +498,10 @@ TEST_CASE_METHOD(
 
     // Submit query
     query.submit();
+    REQUIRE(query.query_status() == Query::Status::COMPLETE);
 
     // The deserialized query should also include the read stats
+    check_subarray_stats(5, 5);
     check_read_stats(query);
 
     auto result_el = query.result_buffer_elements_nullable();
@@ -500,7 +535,9 @@ TEST_CASE_METHOD(
     std::vector<char> a3_data(60);
     std::vector<uint64_t> a3_offsets(4);
     std::vector<int32_t> subarray = {3, 4, 3, 4};
-    query.set_subarray(subarray);
+    Subarray cppapi_subarray(ctx, array);
+    cppapi_subarray.set_subarray(subarray);
+    query.set_subarray(cppapi_subarray);
 
     auto set_buffers = [&](Query& q) {
       q.set_data_buffer("a1", a1);
@@ -515,6 +552,7 @@ TEST_CASE_METHOD(
     query.submit();
     // The deserialized query should also include the read stats
     check_read_stats(query);
+    check_subarray_stats(5, 5);
     REQUIRE(query.query_status() == Query::Status::INCOMPLETE);
 
     auto result_el = query.result_buffer_elements_nullable();
@@ -529,6 +567,7 @@ TEST_CASE_METHOD(
     query.submit();
     // The deserialized query should also include the read stats
     check_read_stats(query);
+    check_subarray_stats(7, 7);
 
     REQUIRE(query.query_status() == Query::Status::INCOMPLETE);
     result_el = query.result_buffer_elements_nullable();
@@ -545,6 +584,7 @@ TEST_CASE_METHOD(
     query.submit();
     // The deserialized query should also include the read stats
     check_read_stats(query);
+    check_subarray_stats(9, 9);
 
     REQUIRE(query.query_status() == Query::Status::COMPLETE);
     result_el = query.result_buffer_elements_nullable();
@@ -575,7 +615,9 @@ TEST_CASE_METHOD(
     std::vector<uint64_t> a3_offsets(1000);
     std::vector<int32_t> subarray = {1, 10, 1, 10};
 
-    query.set_subarray(subarray);
+    Subarray cppapi_subarray(ctx, array);
+    cppapi_subarray.set_subarray(subarray);
+    query.set_subarray(cppapi_subarray);
     query.set_data_buffer("a1", a1);
     query.set_data_buffer("a2", a2);
     query.set_validity_buffer("a2", a2_nullable);
@@ -625,7 +667,9 @@ TEST_CASE_METHOD(
     std::vector<int32_t> subarray = {1, 10, 1, 10};
 
     query.set_layout(TILEDB_GLOBAL_ORDER);
-    query.set_subarray(subarray);
+    Subarray cppapi_subarray(ctx, array);
+    cppapi_subarray.set_subarray(subarray);
+    query.set_subarray(cppapi_subarray);
     query.set_data_buffer("a1", a1);
     query.set_data_buffer("a2", a2);
     query.set_validity_buffer("a2", a2_nullable);
@@ -634,6 +678,7 @@ TEST_CASE_METHOD(
 
     // Submit query
     query.submit();
+    REQUIRE(query.query_status() == Query::Status::COMPLETE);
 
     // The deserialized query should also include the read stats
     check_read_stats(query);
@@ -667,7 +712,9 @@ TEST_CASE_METHOD(
     std::vector<uint64_t> a3_offsets(1000);
     std::vector<int32_t> subarray = {1, 10, 1, 10};
 
-    query.set_subarray(subarray);
+    Subarray cppapi_subarray(ctx, array);
+    cppapi_subarray.set_subarray(subarray);
+    query.set_subarray(cppapi_subarray);
     query.set_coordinates(coords);
     query.set_data_buffer("a1", a1);
     query.set_data_buffer("a2", a2);
@@ -677,6 +724,7 @@ TEST_CASE_METHOD(
 
     // Submit query
     query.submit();
+    REQUIRE(query.query_status() == Query::Status::COMPLETE);
 
     // The deserialized query should also include the read stats
     check_read_stats(query);
@@ -710,10 +758,10 @@ TEST_CASE_METHOD(
     std::vector<uint64_t> a3_offsets(1000);
     std::vector<int32_t> subarray = {1, 10, 1, 10};
 
-    Subarray sub(ctx, array);
-    sub.add_range(0, subarray[0], subarray[1]);
-    sub.add_range(1, subarray[2], subarray[3]);
-    query.set_subarray(sub);
+    Subarray cppapi_subarray(ctx, array);
+    cppapi_subarray.add_range(0, subarray[0], subarray[1]);
+    cppapi_subarray.add_range(1, subarray[2], subarray[3]);
+    query.set_subarray(cppapi_subarray);
     query.set_data_buffer("a1", a1);
     query.set_data_buffer("a2", a2);
     query.set_validity_buffer("a2", a2_nullable);
@@ -722,6 +770,7 @@ TEST_CASE_METHOD(
 
     // Submit query
     query.submit();
+    REQUIRE(query.query_status() == Query::Status::COMPLETE);
 
     // The deserialized query should also include the read stats
     check_read_stats(query);
@@ -746,10 +795,10 @@ TEST_CASE_METHOD(
     std::vector<uint64_t> a3_offsets(1000);
     std::vector<int32_t> subarray = {3, 4, 3, 4};
 
-    Subarray sub(ctx, array);
-    sub.add_range(0, subarray[0], subarray[1]);
-    sub.add_range(1, subarray[2], subarray[3]);
-    query.set_subarray(sub);
+    Subarray cppapi_subarray(ctx, array);
+    cppapi_subarray.add_range(0, subarray[0], subarray[1]);
+    cppapi_subarray.add_range(1, subarray[2], subarray[3]);
+    query.set_subarray(cppapi_subarray);
     query.set_data_buffer("a1", a1);
     query.set_data_buffer("a2", a2);
     query.set_validity_buffer("a2", a2_nullable);
@@ -758,6 +807,7 @@ TEST_CASE_METHOD(
 
     // Submit query
     query.submit();
+    REQUIRE(query.query_status() == Query::Status::COMPLETE);
 
     // The deserialized query should also include the read stats
     check_read_stats(query);
@@ -781,10 +831,10 @@ TEST_CASE_METHOD(
     std::vector<char> a3_data(60);
     std::vector<uint64_t> a3_offsets(4);
     std::vector<int32_t> subarray = {3, 4, 3, 4};
-    Subarray sub(ctx, array);
-    sub.add_range(0, subarray[0], subarray[1]);
-    sub.add_range(1, subarray[2], subarray[3]);
-    query.set_subarray(sub);
+    Subarray cppapi_subarray(ctx, array);
+    cppapi_subarray.add_range(0, subarray[0], subarray[1]);
+    cppapi_subarray.add_range(1, subarray[2], subarray[3]);
+    query.set_subarray(cppapi_subarray);
 
     auto set_buffers = [&](Query& q) {
       q.set_data_buffer("a1", a1);
@@ -796,7 +846,6 @@ TEST_CASE_METHOD(
 
     // Submit initial query.
     set_buffers(query);
-    // Submit query
     query.submit();
     // The deserialized query should also include the read stats
     check_read_stats(query);
@@ -861,7 +910,9 @@ TEST_CASE_METHOD(
     std::vector<uint64_t> a3_offsets(1000);
     std::vector<int32_t> subarray = {1, 10, 1, 10};
 
-    query.set_subarray(subarray);
+    Subarray cppapi_subarray(ctx, array);
+    cppapi_subarray.set_subarray(subarray);
+    query.set_subarray(cppapi_subarray);
     query.set_data_buffer("a1", a1);
     query.set_data_buffer("a2", a2);
     query.set_validity_buffer("a2", a2_nullable);
